Extract SassItem into own file
